Update Basic Ga Api.

+++ ga-20191120/core/Models/GetBasicIpSetResponseBody.cs
@@ -9,65 +9,38 @@ using Tea;
 namespace AlibabaCloud.SDK.Ga20191120.Models
 {
     public class GetBasicIpSetResponseBody : TeaModel {
-        /// <summary>
-        /// 加速地域Id
-        /// </summary>
         [NameInMap("AccelerateRegionId")]
         [Validation(Required=false)]
         public string AccelerateRegionId { get; set; }
 
-        /// <summary>
-        /// 全球加速实例Id
-        /// </summary>
         [NameInMap("AcceleratorId")]
         [Validation(Required=false)]
         public string AcceleratorId { get; set; }
 
-        /// <summary>
-        /// 加速地域带宽
-        /// </summary>
         [NameInMap("Bandwidth")]
         [Validation(Required=false)]
-        public int? Bandwidth { get; set; }
+        public long? Bandwidth { get; set; }
 
-        /// <summary>
-        /// 加速接入点IP地址
-        /// </summary>
         [NameInMap("IpAddress")]
         [Validation(Required=false)]
         public string IpAddress { get; set; }
 
-        /// <summary>
-        /// 加速接入点id
-        /// </summary>
         [NameInMap("IpSetId")]
         [Validation(Required=false)]
         public string IpSetId { get; set; }
 
-        /// <summary>
-        /// 加速接入点地址类型
-        /// </summary>
         [NameInMap("IpVersion")]
         [Validation(Required=false)]
         public string IpVersion { get; set; }
 
-        /// <summary>
-        /// 公网质量类型
-        /// </summary>
         [NameInMap("IspType")]
         [Validation(Required=false)]
         public string IspType { get; set; }
 
-        /// <summary>
-        /// 请求Id
-        /// </summary>
         [NameInMap("RequestId")]
         [Validation(Required=false)]
         public string RequestId { get; set; }
 
-        /// <summary>
-        /// 加速接入点状态
-        /// </summary>
         [NameInMap("State")]
         [Validation(Required=false)]
         public string State { get; set; }
